+++ tests/middleware.ts
@@ -1,0 +1,18 @@
+import { createProcedure } from "../src";
+
+const procedure = createProcedure({
+  createContext: () => {
+    return {
+      ok: true,
+      ctx: {},
+    };
+  },
+});
+
+const verifyMw = procedure.middleware(({ next }) => {
+  return next({ a: 1 });
+});
+
+const testData = procedure.use(verifyMw).query(async ({ ctx }) => {
+  return ctx.a;
+});